functions/csi-ml-service/requirements.txt
flask==2.2.3
numpy==1.24.2
google-cloud-storage==2.7.0
google-cloud-firestore==2.11.0
functions-framework==3.3.0 